refactor: exporting types
